requirements.txt
Django==3.2.16
django-filter==22.1
djangorestframework==3.14.0
Faker==16.6.1
kafka-python==2.0.2
psycopg2-binary==2.9.5
pyspark==3.3.1
pytz==2022.7.1
requests==2.28.2
